requirements.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
bcrypt==4.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.48.2
h5py==3.1.0
idna==3.4
imageio==2.15.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.4.45
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.26.13
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
